requirements.txt
numpy==1.17.4
python-dotenv==0.10.3
